Create error controller and auth middlewares
- Convert error middlewares to a controller.
- Convert ensureAuth to a auth middlewares object.
